Put Command into its own file, refactored it
Command is no longer a sum type because it was only pattern-matched in
`runCommand`. I decided to collapse the sum + runCommand into a single
data type and encapsulate the details. This results in an
object-oriented style: there is a Command interface with a method for
running the command.

One interesting aspect of doing this in Haskell is that a Command can
run in any environment "that has an AlchemyData". The `runCommand`
"method" doesn't take an explicit `AlchemyData`, and the functions
that create Command instances don't either. The Java equivalent is to
have static functions for creating Commands (listIngredients,
listEffectsOf, etc.) and for the Command::run method to accept an
explicit AlchemyData reference.

But since the reference is implicit in Haskell, the Java equivalent in
terms of ergonomics is to have a Command::run method with no arguments
and to create Commands using methods on an AlchemyCommandFactory which
receives an AlchemyData in its constructor.

The factory approach could be implemented in Haskell too. Since both
approaches are possible in both languages, I wonder which is "better".